+++ src/components/Search/index.js
@@ -8,6 +8,7 @@ import { styled } from "@mui/material/styles"
 
 // components
 import SortTool from "./SortTool"
+import Drawer from "./Drawer.js"
 
 const sortList = [
   {
@@ -53,6 +54,7 @@ export default function Search() {
 
   return (
     <React.Fragment key="search">
+      <Drawer />
       <SearchButton onClick={handleOpenSearch}>
         <SearchIcon sx={{ fontSize: "30px", color: "#ffffff" }} />
       </SearchButton>
